# (slightly) faster imgui rendering.