@einarf 2020-04-06
Thanks a lot for this. Looks like this can make a huge difference actually. That little extra time per frame is gold. I'll play a bit with this on all 3 platforms to confirm the result.

We can do a patch release tomorrow if this would practical.