@einarf 2020-04-19
New version released: https://pypi.org/project/moderngl-window/2.1.1/